bot/requirements.txt
aiogram==3.10.0
sqlalchemy==2.0.25
apscheduler==3.10.4
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0